% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/6_BN_clustering.R
\name{BN_clustering}
\alias{BN_clustering}
\title{Clustering Gaussian Bayesian network}
\usage{
BN_clustering(bayesian_structure, data, clusters, initialization, verbose)
}
\arguments{
\item{bayesian_structure}{a bn object from bnlearn package whose nodes are gaussians}

\item{data}{the same dataset used to compute the bayesian_structure}

\item{clusters}{array of values where each value indicates the number of clusters for that model}

\item{initialization}{a natural number to initialize the algorithm. Use 0 to use kmeans initialization and any other positive number as a seed for a random initialization}

\item{verbose}{show BIC scores for each model and the optimal number of clusters}
}
\value{
model_list is the list of parameters that best fit the data according to BIC criteria
}
\description{
Given a Bayesian network structure, apply clustering to the data with the aim to find clusters of somas
}
\examples{
somas <- read.table(file.path(tempdir(),"somaReebParameters.csv"), sep = ";", header = T, row.names = 1)
bayesian_structure <- BN_structure_learning(path_to_csv = file.path(tempdir(),"somaReebParameters.csv"), nboots = 200, significant_arcs = 0.95)
fittest_model <- BN_clustering(bayesian_structure, data = somas, clusters = c(2, 3, 4), initialization = 2, T)
}

