% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/5_BN_structure_learning.R
\name{BN_structure_learning}
\alias{BN_structure_learning}
\title{Compute the structure of the BN}
\usage{
BN_structure_learning(somas, nboots, significant_arcs)
}
\arguments{
\item{somas}{dataframe with the characterization of the somas}

\item{significant_arcs}{a value between 0 and 1 that denotes a threshold over percentage of times that an arc appeared in the boostrap structures}

\item{n_boot}{number of times that bootstrap learn the bayesian structure}
}
\value{
bayesian_model an object of class bn from bnlearn package that representates the structure of the DSBN
}
\description{
Compute the structure of a Dynamical Spatial Bayesian Network. Also apply bootstrap to choose the most significant arcs.
}

