% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/7_simulation.R
\name{cluster_simulation}
\alias{cluster_simulation}
\title{Simulation of new somas from a selected cluster}
\usage{
cluster_simulation(model, data, cluster_number, number_of_new_instances,
  seed = 1)
}
\arguments{
\item{model}{a bn.fit object from bnlearn package whose nodes are gaussians. It must be the result of the BN_clustering function}

\item{data}{the same dataset used to compute the bayesian_structure}

\item{cluster_number}{a natural number between 1 and the number of clusters of model. Simulate new somas from the selected cluster.}

\item{number_of_new_instances}{a natural number denoting the number of new somas sampled from simulation}

\item{seed}{a natural number used as seed during the sampling, by default it is 1}
}
\value{
new_somas a dataset similar to data with the simulated somas of the selected cluster
}
\description{
Simulation of new somas from the choosen cluster of the clustering of bayesian networks
}

