% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3_validation.R
\name{compute_meshes_volumes}
\alias{compute_meshes_volumes}
\title{Compute the volume of meshes in a folder}
\usage{
compute_meshes_volumes(path_to_mesh_folder)
}
\arguments{
\item{path_to_somas_folder}{path to the folder where meshes processed with the first technique are placed. They have to be PLY files}
}
\value{
array of N volumes where N is the number of meshes in the folder
}
\description{
Read one by one the meshes in a folder and compute the volume of each one of the them
}
\examples{
compute_meshes_volumes("./temp/final_result")
}

