% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3_validation.R
\name{MAQ_S}
\alias{MAQ_S}
\title{Compute MAQ_S}
\usage{
MAQ_S(path_somas_algorithm, path_somas_expert_1, path_somas_expert_2)
}
\arguments{
\item{path_somas_algorithm}{path to the folder where meshes processed with the first technique are placed. They have to be PLY files}

\item{path_somas_expert_1}{path to the folder where meshes processed with the second technique are placed. They have to be PLY files}

\item{path_somas_expert_2}{path to the folder where meshes processed with the third technique are placed. They have to be PLY files}
}
\value{
A vector of three values where each value is the MAQ_S between two techniques. Concretely they are MAQ_S_12, MAQ_S_13, MAQ_S_23
}
\description{
Compute MAQ_S between three different techniques to compare processing methods
}
\examples{
path_somas_algorithm<- "./temp/final_result"
path_somas_experts<- system.file("test/post_repaired",package="SomaMS")
experts_paths<-list.dirs(path_somas_experts,recursive=F)
path_somas_expert_1<-experts_paths[1]
path_somas_expert_2<-experts_paths[2]
MAQ_S_result <- MAQ_S(path_somas_algorithm,path_somas_expert_1, path_somas_expert_2)
print(paste("MAQ_S_12 is:",   MAQ_S_result[1] * 100, "\%"))
print(paste("MAQ_S_13 is:",   MAQ_S_result[2] * 100, "\%"))
print(paste("MAQ_S_23 is:",   MAQ_S_result[3] * 100, "\%"))
print(paste("Mean MAQ_S between algorithm and experts is:", mean(c(MAQ_S_result[1],MAQ_S_result[2])) * 100, "\%"))
print(paste("Difference between experts MAQ_S and mean MAQ_S of algorithm is:", abs(MAQ_S_result[3]-mean(c(MAQ_S_result[1],MAQ_S_result[2])))*100, "\%"))
}

