% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IO-mesh.R
\name{read_binary_PLY}
\alias{read_binary_PLY}
\title{Read binary PLY file}
\usage{
read_binary_PLY(filePath)
}
\arguments{
\item{file_name}{path to the PLY file}
}
\value{
a list of three components of the mesh, the vertices, the faces and the color or quality of the vertices
}
\description{
Read vertices, faces and color or quality of the vertices of a mesh from a PLY file in little endian 1.0 format
}

