% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IO-mesh.R
\name{read_VRML}
\alias{read_VRML}
\title{Read neuron from a VRML file}
\usage{
read_VRML(file_name)
}
\arguments{
\item{file_name}{path to the PLY file}
}
\value{
a list of two components of the mesh, the vertices and the faces
}
\description{
Read vertices, faces of a mesh from a VRML file. Adapted from package geomorph
}

