% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_repair_soma.R
\name{repair.soma3d}
\alias{repair.soma3d}
\title{Repair single soma from data}
\usage{
repair.soma3d(soma, ao.path = tempdir(), broken.path = tempdir(),
  repaired.path = tempdir(), name = "soma", returnSoma = F)
}
\arguments{
\item{soma}{A filepath or the structure returned by read_VRML or read_PLY representing damaged soma mesh}

\item{ao.path}{folder where intermediate files with ambient occlusion data are saved}

\item{broken.path}{Folder where intermediate files with broken somas after GMM threshold are saved}

\item{repaired.path}{Folder where final file is stored}

\item{name}{Name of the final file}

\item{returnSoma}{}
}
\value{
Repaired soma (if returnSoma is True)
}
\description{
Repair the surface of the soma to remove holes and vertices inside the mesh
}
\seealso{
\code{read_VRML}
}

