% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_repair_soma.R
\name{repair_soma}
\alias{repair_soma}
\title{Repair soma morphology}
\usage{
repair_soma(directory, output_ambient_occlusion, broken_mesh,
  output_poisson_reconstruction)
}
\arguments{
\item{directory}{path to the folder where PLY files representing damaged soma mesh}

\item{output_ambient_occlusion}{path to the folder where somas with ambient occlusion are saved}

\item{broken_mesh}{path to the folder where somas were saved after the GMM threshold was imposed and vertices inside the soma or forming cavities were removed}

\item{output_poisson_reconstruction}{path to the folder where somas were saved after the mesh closing process}
}
\value{
None
}
\description{
Repair the surface of the neuron to remove holes and vertices inside the mesh
}
\examples{
repair_soma(directory = file.path(tempdir(),"PLYs"), output_ambient_occlusion = file.path(tempdir(),"ambient_occlusion"), broken_mesh = file.path(tempdir(),"broken_mesh_ao"), output_poisson_reconstruction = file.path(tempdir(),"poisson_reconstruction_ao"))
}

