% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2_extract_soma.R
\name{segment_somas}
\alias{segment_somas}
\title{Segment somas}
\usage{
segment_somas(directory, output_shape_diameter, broken_mesh,
  output_poisson_reconstruction, final_result)
}
\arguments{
\item{directory}{path to the folder where PLY files of the somas are saved}

\item{output_shape_diameter}{path to the folder where somas with shape diameter will be saved}

\item{broken_mesh}{path to the folder where somas will be saved after that the GMM threshold removes the dendrites}

\item{output_poisson_reconstruction}{path to the folder where somas will be saved after the mesh closing process}

\item{final_result}{path to the folder where final segmented somas will be saved after the isolated pieces are removed}
}
\value{
None
}
\description{
Apply Gaussian mixture models to segment soma from dendrites
}
\examples{
segment_somas(directory = file.path(tempdir(), "poisson_reconstruction_ao"), output_shape_diameter = file.path(tempdir(), "sdf"), broken_mesh = file.path(tempdir(), "broken_mesh_sdf"), output_poisson_reconstruction = file.path(tempdir(), "poisson_reconstruction_sdf"), final_result = file.path(tempdir(), "final_result"))
}

