% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/7_simulation.R
\name{simulate_new_somas}
\alias{simulate_new_somas}
\title{Simulation of new somas}
\usage{
simulate_new_somas(model, data, cluster_number = NULL,
  number_of_new_instances = 100, seed = 1)
}
\arguments{
\item{model}{a bn.fit object from bnlearn package whose nodes are gaussians. It must be the result of the BN_clustering function}

\item{data}{the same dataset used to compute the bayesian_structure}

\item{cluster_number}{a number between 1 and the number of clusters of model. When it is NULL, new somas are generated from the priori distribution. When is a number between 1 and the number of clusters, simulate new somas from the selected cluster.}

\item{number_of_new_instances}{a natural number denoting the number of new somas sampled from simulation}

\item{seed}{a natural number used as seed during the sampling, by default it is 1}
}
\value{
new_somas a dataset similar to data with the simulated somas
}
\description{
Simulation of new somas of the clustering of bayesian networks
}

