% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/7_simulation.R
\name{simulation_3D_mesh}
\alias{simulation_3D_mesh}
\title{Three-dimensional representation of the simulated somas}
\usage{
simulation_3D_mesh(new_somas, index)
}
\arguments{
\item{new_somas}{a dataset with the simulated somas obtained in simulate_new_somas function}

\item{index}{index of the row of the new soma to renderize}
}
\value{
curve_points a n x 3 dataframe where each row represents a point in the cartesian three-dimensional space
}
\description{
Three-dimensional renderization from a dataset where the simulated somas were saved
}

