% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4_soma_parametrization.R
\name{soma_caracterization}
\alias{soma_caracterization}
\title{Characterize soma with ray tracing reeb graph}
\usage{
soma_caracterization(path_somas_repaired, path_csv_file)
}
\arguments{
\item{path_somas_repaired}{path to the folder where segmented somas were placed. They have to be PLY files.}

\item{path_csv_file}{path to a csv file where the values of the characterization will be saved}
}
\value{
None
}
\description{
Compute the morphological characterization of the soma. First, cast rays from the centroid of the soma to the surface,
defining level curves. Next, orientate soma and adjust curves. Finally compute parameters according to the curves.
}
\examples{
soma_caracterization(path_somas_repaired = file.path(tempdir(),"reconstructed"), path_csv_file = file.path(tempdir(),"somaReebParameters.csv"))
}

