% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3_validation.R
\name{wilcoxon_RMSE}
\alias{wilcoxon_RMSE}
\title{Wilcoxon test for RMSE}
\usage{
wilcoxon_RMSE(RMSE_matrix)
}
\arguments{
\item{RMSE_matrix}{a matrix Nx3 which is the output of}
}
\value{
A vector of two p-values. The first value is the p-value resulting from wilcoxon between first column of the RMSE matrix and second column of the RMSE matrix. The second one is the p-value resulting from wilcoxon between first column of the RMSE matrix and second column of the RMSE matrix
}
\description{
Compute Wilcoxon between RMSE distances to find significant differences between mesh processing techniques
}
\examples{
path_somas_algorithm<- "./test/final_result"
path_somas_experts<- system.file("test/pre_repaired",package="SomaMS")
experts_paths<-list.dirs(path_somas_experts,recursive=F)
pre_repaired_RMSE<-RMSE_mesh_distance(path_somas_algorithm, experts_paths[1], experts_paths[2], TRUE)

#RMSE between the somas of the algorithm and the somas of the experts after repairing the experts' somas
path_somas_algorithm<- "./test/final_result"
path_somas_experts<- system.file("test/post_repaired",package="SomaMS")
experts_paths<-list.dirs(path_somas_experts,recursive=F)
post_repaired_RMSE<-RMSE_mesh_distance(path_somas_algorithm, experts_paths[1], experts_paths[2], TRUE)

pvalue_before_repairing <- wilcoxon_RMSE(pre_repaired_RMSE)
print(paste0("p-value between algorithm and first expert: ", pvalue_before_repairing[1]))
print(paste0("p-value between algorithm and second expert: ", pvalue_before_repairing[2]))

pvalue_after_repairing <- wilcoxon_RMSE(post_repaired_RMSE)
print(paste0("p-value between algorithm and first expert: ", pvalue_after_repairing[1]))
print(paste0("p-value between algorithm and second expert: ", pvalue_after_repairing[2]))
}

