% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IO-mesh.R
\name{write_PLY_only_points}
\alias{write_PLY_only_points}
\title{Write only the vertices of a mesh in a PLY file}
\usage{
write_PLY_only_points(vertices, file_name = "")
}
\arguments{
\item{vertices}{matrix Nx3 where each row represents the coordinates of a vertex of the mesh}

\item{file_name}{path and name of the PLY file where point cloud is going to be saved}
}
\value{
None
}
\description{
Write only the vertices of a mesh in a PLY file so it is saved a point cloud
}

